second_app/serializers.py
from rest_framework import serializers

from second_app.models import University


class UniversitySerializer(serializers.ModelSerializer):
    class Meta:
        model = University
        fields = '__all__'
