requirements.txt
asgiref==3.7.2
dj-rest-auth==4.0.1
Django==4.2.2
djangorestframework==3.14.0
drf-yasg==1.21.6
et-xmlfile==1.1.0
inflection==0.5.1
numpy==1.25.0
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
